update GDAL to 3.10.0 (#183)
* upgrade GDAL to 3.10.0

* upgrade Node.js platforms

* conform to the new stricter afterEach signature

* update the changelog

* gti is a separate driver now

* fix mitab

* fix internal shape

* gmutils

* files moved to root ogr

* add the new files

* test leaving the C++ standard to node-gyp

* try the magic MSVC option

* fix the not so orthodox code

* try C++17 only for PROJ

* reference directly in the target

* add a compilertool section

* remove the verbose build

* switch to macos-13

+++ deps/libgdal/gdal/ogr/ogrsf_frmts/mitab/mitab_indfile.cpp
@@ -11,23 +11,7 @@
  * Copyright (c) 1999-2001, Daniel Morissette
  * Copyright (c) 2014, Even Rouault <even.rouault at spatialys.com>
  *
- * Permission is hereby granted, free of charge, to any person obtaining a
- * copy of this software and associated documentation files (the "Software"),
- * to deal in the Software without restriction, including without limitation
- * the rights to use, copy, modify, merge, publish, distribute, sublicense,
- * and/or sell copies of the Software, and to permit persons to whom the
- * Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included
- * in all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.  IN NO EVENT SHALL
- * THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
- * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
- * DEALINGS IN THE SOFTWARE.
+ * SPDX-License-Identifier: MIT
  **********************************************************************/
 
 #include "cpl_port.h"
@@ -1788,7 +1772,7 @@ int TABINDNode::SplitNode()
 #ifdef DEBUG
         // Just in case, reset space previously used by moved entries
         memset(m_poDataBlock->GetCurDataPtr(), 0,
-               numInNode2 * (m_nKeyLength + 4));
+               static_cast<size_t>(numInNode2) * (m_nKeyLength + 4));
 #endif
         // And update current node members
         m_numEntriesInNode = numInNode1;
@@ -1860,7 +1844,7 @@ int TABINDNode::SplitNode()
         // Just in case, reset space previously used by moved entries
         memset(m_poDataBlock->GetCurDataPtr() +
                    static_cast<size_t>(numInNode2) * (m_nKeyLength + 4),
-               0, numInNode1 * (m_nKeyLength + 4));
+               0, static_cast<size_t>(numInNode1) * (m_nKeyLength + 4));
 #endif
 
         // And update current node members
@@ -1943,7 +1927,7 @@ int TABINDNode::SplitRootNode()
 #ifdef DEBUG
     // Just in case, reset space previously used by moved entries
     memset(m_poDataBlock->GetCurDataPtr(), 0,
-           m_numEntriesInNode * (m_nKeyLength + 4));
+           static_cast<size_t>(m_numEntriesInNode) * (m_nKeyLength + 4));
 #endif
 
     /*-----------------------------------------------------------------
